pyproject.toml
[project]
name = "looker-report-explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-aiplatform>=1.133.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
